Okay, after proper analysis, I have figured out what the problem is.
Whenever I upload a document larger than 500kb, it doesn't work and it show me this error message:
Error: An unexpected response was received from the server.
    at fetchServerAction (https://6000-firebase-studio-1747620016344.cluster-ombtxv25tbd6yrjpp3lukp6zhc.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_client_43e3ffb8._.js:17725:37)
    at HomePage (rsc://React/Server/file:///home/user/studio/.next/server/chunks/ssr/%5Broot%20of%20the%20server%5D__9cc4a647._.js?23:703:278)

Now can you change that to allow the user to upload documents as large as 10mb max and let the user know the maximum amount they can upload

Also, can you change the way the results are outputted. This is a site that is suppose to help school students with their school work; I am satisfied with the summary result output but can we adjust the other output results

For "question answering": can the output be a series of questions generated from the file that the user uploaded together with answers under it like this:
> Question: What is Object Oriented Programming?"
> Answer: Object Oriented Programming or OOP is ......

> Question: What is the 4 pillars of Object Oriented Programming?"
> Answer: Encapsulation, abstraction, .....

This will allow the user to use our site to prepare for exams and all. It will be a good tool.

For "Text": I am satisfied with the result output but please note that the text tool is for asssignments and it is used to solve assignments questions and produce a document result of the solution. This will be used for courses that require a lot of research and a lot of students will use it