The output result of the explain topic/question looks better than it looked before, thank you for that
It still looks different from the generate q&a output results, maybe if you can make the headers and sub headers to be a different color or something just like the generate q&a are

Also, can you put a stationary text field bar under the result bar that doesn't move. So the user can ask more follow up questions if they want to. Then out a download icon and a schedule icon inside the text field bar for a better user interface. The text field bar will be stationary on the screen, the bottom section of the result screen

Change the title of the result screen from "Result!" to "Study Results..."