The error message that reads No results found, let's tweak the message
Tweak the message to:

It seems there's no study results to display. Please try generating one first

Go to Study

I have seen the changes made to the results page and I can say that we are making progress. Now can we make the text field bar to be elevated on the screen, I want tot be able to see my right reserved etc and the text field bar shouldn't span across the screen length and it should have a send button at the right side of the text field bar